redcloud-inference/core/requirements.txt

Pillow
pydantic==2.9.2
ultralytics
opencv-contrib-python